+++ tsconfig.json
@@ -20,5 +20,5 @@
     "noEmit": true,
     "jsx": "react-jsx",
   },
-  "include": ["src/**/*"],
+  "include": ["src/**/*", "helper/**/*"],
 }
